refactor: move charm plugin to its own file
